Gateway GZ80x: add dtsi and power off support via pwr button
There are three revs: AM-GZ80x, AM-GZ80x-US and SC-GZ80x

Add a DSTI in case one should drop out of the sky and we
need a DTS for it. This also removes its dependency on
meson-axg-jethome-jethub-j1xx.dtsi.

Can now power down via power button on the back of the
unit and power back on via reset button on the side.

Signed-off-by: Patrick Yavitz <pyavitz@gmail.com>