rk35xx-vendor: build DRM_PANEL_MIPI_DBI as builtin
This driver will request firmware by the first compatiable string from
devicetree, which is usually a device specific string not in the driver
code. If this driver is built as module, it can't get loaded because
only the first compatible string is used to probe driver. There is no
such issue if this driver is built in kernel.